moved create account functionality to controller
